break out get_projects_for_obj
